tests/test_blynk_integration.py
# tests/test_blynk_integration.py

import unittest
from unittest.mock import patch, MagicMock
from sim7020py.blynk_integration import BlynkIntegration
from sim7020py.commands import ATCommandError


class TestBlynkIntegration(unittest.TestCase):

    @patch('serial.Serial')
    def setUp(self, mock_serial):
        """
        Set up the BlynkIntegration instance with a mock serial connection for testing.
        """
        self.mock_serial = mock_serial.return_value
        self.blynk = BlynkIntegration(port="COM_TEST", apn="test_apn", blynk_token="test_token")

    def test_connect_success(self):
        """
        Test that connect establishes network connection and initializes Blynk.
        """
        self.mock_serial.readlines.return_value = [b"OK\r\n"]

        # Call connect and ensure no exceptions are raised
        self.blynk.connect()

        # Verify that the APN setup and network connection commands were sent
        self.mock_serial.write.assert_any_call(b'AT+CGDCONT=1,"IP","test_apn"\r\n')
        self.mock_serial.write.assert_any_call(b"AT+CGATT=1\r\n")

    def test_connect_failure(self):
        """
        Test that connect raises ATCommandError if the network connection fails.
        """
        self.mock_serial.readlines.return_value = [b"ERROR\r\n"]

        # Expect an ATCommandError to be raised on connect
        with self.assertRaises(ATCommandError):
            self.blynk.connect()

    def test_send_value_success(self):
        """
        Test that send_value sends data to the correct Blynk virtual pin.
        """
        self.mock_serial.readlines.return_value = [b"OK\r\n"]

        virtual_pin = 1
        value = 25
        self.blynk.send_value(virtual_pin, value)

        expected_command = (
            b'AT+HTTPGET="http://blynk-cloud.com/test_token/update/1?value=25"\r\n'
        )
        self.mock_serial.write.assert_called_with(expected_command)

    def test_send_value_failure(self):
        """
        Test that send_value handles failure when unable to send data.
        """
        self.mock_serial.readlines.return_value = [b"ERROR\r\n"]

        virtual_pin = 1
        value = 25
        with self.assertRaises(ATCommandError):
            self.blynk.send_value(virtual_pin, value)

    def test_get_value_success(self):
        """
        Test that get_value retrieves data from the correct Blynk virtual pin.
        """
        self.mock_serial.readlines.return_value = [b'"25"\r\n', b"OK\r\n"]

        virtual_pin = 1
        result = self.blynk.get_value(virtual_pin)

        self.assertEqual(result, "25")

        expected_command = (
            b'AT+HTTPGET="http://blynk-cloud.com/test_token/get/1"\r\n'
        )
        self.mock_serial.write.assert_called_with(expected_command)

    def test_get_value_failure(self):
        """
        Test that get_value handles failure when unable to retrieve data.
        """
        self.mock_serial.readlines.return_value = [b"ERROR\r\n"]

        virtual_pin = 1
        result = self.blynk.get_value(virtual_pin)
        self.assertIsNone(result)

    def test_disconnect(self):
        """
        Test that disconnect sends the correct command to detach from the network.
        """
        self.mock_serial.readlines.return_value = [b"OK\r\n"]

        self.blynk.disconnect()
        self.mock_serial.write.assert_called_with(b"AT+CGATT=0\r\n")

    def tearDown(self):
        """
        Clean up after each test.
        """
        self.blynk.close()
        self.mock_serial.close.assert_called_once()


if __name__ == "__main__":
    unittest.main()

# import unittest
# from unittest.mock import patch, MagicMock
# from sim7020py.blynk_integration import BlynkIntegration
# from sim7020py.commands import ATCommandError
#
#
# class TestBlynkIntegration(unittest.TestCase):
#
#     @patch('serial.Serial')
#     def setUp(self, mock_serial):
#         """
#         Set up the BlynkIntegration instance with a mock serial connection for testing.
#         """
#         self.mock_serial = mock_serial.return_value
#         self.blynk = BlynkIntegration(port="COM_TEST", apn="test_apn", blynk_token="test_token")
#
#     def test_connect_success(self):
#         """
#         Test that connect establishes network connection and initializes Blynk.
#         """
#         # Mock responses for network and server setup
#         self.mock_serial.readlines.return_value = [b"OK\r\n"]
#
#         # Call connect and ensure no exceptions are raised
#         self.blynk.connect()
#
#         # Verify that the APN setup and network connection commands were sent
#         self.mock_serial.write.assert_any_call(b'AT+CGDCONT=1,"IP","test_apn"\r\n')
#         self.mock_serial.write.assert_any_call(b"AT+CGATT=1\r\n")
#
#     def test_connect_failure(self):
#         """
#         Test that connect raises ATCommandError if the network connection fails.
#         """
#         # Mock response for a failed connection attempt
#         self.mock_serial.readlines.return_value = [b"ERROR\r\n"]
#
#         # Expect an ATCommandError to be raised on connect
#         with self.assertRaises(ATCommandError):
#             self.blynk.connect()
#
#     def test_send_value_success(self):
#         """
#         Test that send_value sends data to the correct Blynk virtual pin.
#         """
#         # Mock response to simulate successful data send
#         self.mock_serial.readlines.return_value = [b"OK\r\n"]
#
#         # Call send_value with a test virtual pin and value
#         virtual_pin = 1
#         value = 25
#         self.blynk.send_value(virtual_pin, value)
#
#         # Check that the correct HTTP GET command was sent
#         expected_command = (
#             b'AT+HTTPGET="http://blynk-cloud.com/test_token/update/1?value=25"\r\n'
#         )
#         self.mock_serial.write.assert_called_with(expected_command)
#
#     def test_send_value_failure(self):
#         """
#         Test that send_value handles failure when unable to send data.
#         """
#         # Mock response for a failed send attempt
#         self.mock_serial.readlines.return_value = [b"ERROR\r\n"]
#
#         # Expect an ATCommandError to be raised
#         virtual_pin = 1
#         value = 25
#         with self.assertRaises(ATCommandError):
#             self.blynk.send_value(virtual_pin, value)
#
#     def test_get_value_success(self):
#         """
#         Test that get_value retrieves data from the correct Blynk virtual pin.
#         """
#         # Mock response to simulate successful data retrieval
#         self.mock_serial.readlines.return_value = [b'"25"\r\n', b"OK\r\n"]
#
#         # Call get_value with a test virtual pin
#         virtual_pin = 1
#         result = self.blynk.get_value(virtual_pin)
#
#         # Check the retrieved value
#         self.assertEqual(result, "25")
#
#         # Check that the correct HTTP GET command was sent
#         expected_command = (
#             b'AT+HTTPGET="http://blynk-cloud.com/test_token/get/1"\r\n'
#         )
#         self.mock_serial.write.assert_called_with(expected_command)
#
#     def test_get_value_failure(self):
#         """
#         Test that get_value handles failure when unable to retrieve data.
#         """
#         # Mock response for a failed retrieval attempt
#         self.mock_serial.readlines.return_value = [b"ERROR\r\n"]
#
#         # Call get_value and expect None due to failure
#         virtual_pin = 1
#         result = self.blynk.get_value(virtual_pin)
#         self.assertIsNone(result)
#
#     def test_disconnect(self):
#         """
#         Test that disconnect sends the correct command to detach from the network.
#         """
#         # Mock response for a successful disconnect
#         self.mock_serial.readlines.return_value = [b"OK\r\n"]
#
#         # Call disconnect
#         self.blynk.disconnect()
#
#         # Check that the correct command was sent
#         self.mock_serial.write.assert_called_with(b"AT+CGATT=0\r\n")
#
#     def tearDown(self):
#         """
#         Clean up after each test.
#         """
#         self.blynk.close()
#         self.mock_serial.close.assert_called_once()
#
#
# if __name__ == "__main__":
#     unittest.main()





# import unittest
# from unittest.mock import MagicMock, patch
# from sim7020py.blynk_integration import BlynkIntegration  # Adjust the import as necessary
# import time
#
# class TestBlynkIntegration(unittest.TestCase):
#
#     @patch('sim7020py.blynk_integration.SIM7020')  # Mock SIM7020
#     def setUp(self, MockSIM7020):
#         # Create a mock SIM7020 object
#         self.mock_sim7020 = MagicMock()
#         MockSIM7020.return_value = self.mock_sim7020
#
#         # Create an instance of BlynkIntegration
#         self.blynk = BlynkIntegration(port="/dev/ttyUSB0", apn="internet", blynk_token="test_token")
#
#     def test_initialization(self):
#         # Check that SIM7020 is initialized with the given parameters
#         self.assertEqual(self.blynk.sim7020, self.mock_sim7020)
#         self.assertEqual(self.blynk.apn, "internet")
#         self.assertEqual(self.blynk.blynk_token, "test_token")
#
#     @patch('time.sleep', return_value=None)  # Mock time.sleep to avoid delays
#     def test_connect_success(self):
#         # Mock the SIM7020 methods to simulate a successful connection
#         self.mock_sim7020.initialize.return_value = None
#         self.mock_sim7020.set_apn.return_value = None
#         self.mock_sim7020.connect_network.return_value = None
#
#         # Call the connect method
#         self.blynk.connect()
#
#         # Check if the connection methods were called
#         self.mock_sim7020.initialize.assert_called_once()
#         self.mock_sim7020.set_apn.assert_called_once_with("internet")
#         self.mock_sim7020.connect_network.assert_called_once()
#
#         # Check if the connection status is updated
#         self.assertTrue(self.blynk.connected)
#
# @patch('time.sleep', return_value=None)
# def test_connect_failure(self):
#     # Mock the SIM7020 methods to simulate a connection failure
#     self.mock_sim7020.initialize.side_effect = Exception("Connection failed")
#
#     # Call the connect method
#     self.blynk.connect()
#
#     # Check if the connection status is not updated
#     self.assertFalse(self.blynk.connected)
#
# @patch('time.sleep', return_value=None)
# def test_connect_timeout(self):
#     # Simulate a timeout during connection
#     self.mock_sim7020.connect_network.side_effect = TimeoutError("Connection timeout")
#
#     # Call the connect method
#     self.blynk.connect()
#
#     # Check if the connection status is not updated
#     self.assertFalse(self.blynk.connected)
#     self.mock_sim7020.connect_network.assert_called_once()
#
# def test_send_value_success(self):
#     # Mock the send_command method to simulate a successful value send
#     self.mock_sim7020.at_command.send_command.return_value = ["OK"]
#
#     # Call the send_value method
#     self.blynk.send_value(virtual_pin=1, value=100)
#
#     # Check if the send_command was called with the correct command
#     self.mock_sim7020.at_command.send_command.assert_called_with(
#         'AT+HTTPGET="http://{}/test_token/update/1?value=100"'.format(self.blynk.blynk_server_ip),
#         expected_response="OK"
#     )
#
# def test_send_value_failure(self):
#     # Mock the send_command method to simulate a failure in sending the value
#     self.mock_sim7020.at_command.send_command.side_effect = Exception("Send failed")
#
#     # Call the send_value method
#     self.blynk.send_value(virtual_pin=1, value=100)
#
#     # Check if the send_command method was retried (based on max_retries)
#     self.assertEqual(self.mock_sim7020.at_command.send_command.call_count, self.blynk.max_retries)
#
# def test_send_value_malformed_response(self):
#     # Simulate a malformed response
#     self.mock_sim7020.at_command.send_command.return_value = ["ERROR"]
#
#     # Call the send_value method
#     self.blynk.send_value(virtual_pin=1, value=100)
#
#     # Check if the send_command was retried after receiving an unexpected response
#     self.assertEqual(self.mock_sim7020.at_command.send_command.call_count, self.blynk.max_retries)
#
# def test_get_value_success(self):
#     # Mock the send_command method to simulate a successful value retrieval
#     self.mock_sim7020.at_command.send_command.return_value = ["OK", "value=100"]
#
#     # Call the get_value method
#     value = self.blynk.get_value(virtual_pin=1)
#
#     # Check if the response was parsed correctly
#     self.assertEqual(value, "100")
#
# def test_get_value_failure(self):
#     # Mock the send_command method to simulate a failure in retrieving the value
#     self.mock_sim7020.at_command.send_command.side_effect = Exception("Get failed")
#
#     # Call the get_value method
#     value = self.blynk.get_value(virtual_pin=1)
#
#     # Check if None was returned after retries
#     self.assertIsNone(value)
#
# def test_get_value_malformed_response(self):
#     # Simulate a malformed response (e.g., no value in response)
#     self.mock_sim7020.at_command.send_command.return_value = ["OK", "malformed_response"]
#
#     # Call the get_value method
#     value = self.blynk.get_value(virtual_pin=1)
#
#     # Ensure the return value is None due to incorrect response
#     self.assertIsNone(value)
#
# @patch('time.sleep', return_value=None)
# def test_disconnect(self):
#     # Call the disconnect method
#     self.blynk.disconnect()
#
#     # Check if the disconnect method of SIM7020 was called
#     self.mock_sim7020.disconnect_network.assert_called_once()
#     self.assertFalse(self.blynk.connected)
#
# @patch('time.sleep', return_value=None)
# def test_disconnect_failure(self):
#     # Simulate a failure in disconnecting
#     self.mock_sim7020.disconnect_network.side_effect = Exception("Disconnect failed")
#
#     # Call the disconnect method
#     self.blynk.disconnect()
#
#     # Ensure the method was called, even though it failed
#     self.mock_sim7020.disconnect_network.assert_called_once()
#
# def test_close(self):
#     # Call the close method
#     self.blynk.close()
#
#     # Check if the close method of SIM7020 was called
#     self.mock_sim7020.close.assert_called_once()
#
# def test_max_retries(self):
#     # Test that the max_retries limit is respected
#     self.mock_sim7020.at_command.send_command.side_effect = Exception("Command failed")
#
#     # Call the send_value method
#     self.blynk.send_value(virtual_pin=1, value=100)
#
#     # Ensure that the method is retried the maximum number of times
#     self.assertEqual(self.mock_sim7020.at_command.send_command.call_count, self.blynk.max_retries)
#
# def test_empty_apn(self):
#     # Test with empty APN string
#     blynk_empty_apn = BlynkIntegration(port="/dev/ttyUSB0", apn="", blynk_token="test_token")
#     with self.assertRaises(ValueError):
#         blynk_empty_apn.connect()
#
# def test_empty_token(self):
#     # Test with empty Blynk token
#     blynk_empty_token = BlynkIntegration(port="/dev/ttyUSB0", apn="internet", blynk_token="")
#     with self.assertRaises(ValueError):
#         blynk_empty_token.connect()
#
# if __name__ == '__main__':
#     unittest.main()